0.3.0
refactoring
