+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.github.Krumuvecis</groupId>
     <artifactId>RotorCalculator</artifactId>
-    <version>0.1</version>
+    <version>0.3.0</version>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
