--- src/main/java/rotorCalculator/data/rotorModel/limits/StructuralLimits.java
@@ -1,6 +1,0 @@
-package rotorCalculator.data.rotorModel.limits;
-
-// TODO: add javadoc
-public class StructuralLimits {
-    // TODO: finish this
-}
