+++ src/main/java/rotorCalculator/rotorModel/limits/StructuralLimits.java
@@ -1,0 +1,6 @@
+package rotorCalculator.rotorModel.limits;
+
+// TODO: add javadoc
+public class StructuralLimits {
+    // TODO: finish this
+}
