+++ src/test/java/graphicalOutput/graphics/panels/leftPanel/Column2.java
@@ -1,0 +1,21 @@
+package graphicalOutput.graphics.panels.leftPanel;
+
+import graphicsEngine.panels.PanelColors;
+import org.jetbrains.annotations.Nullable;
+import graphicalOutput.graphics.panels.leftPanel.sections.ParticularSections;
+
+import javax.swing.*;
+
+public class Column2 extends AbstractColumn {
+    public static final int WIDTH = 250;
+
+    protected Column2(@Nullable PanelColors panelColors) {
+        super(WIDTH, panelColors, true);
+    }
+
+    @Override
+    public void addParts(JPanel p) {
+        PanelColors colors = getPanelColors();
+        p.add(new ParticularSections.ActualParameterSection(colors));
+    }
+}
